Rework "open storage" action
The previous implementation was a deeply nested decision tree for
different error cases or other unexpected states.

Most of this was necessary, because setting up correct permissions is partially
happening asynchronously. Some of the redirects would eventually have rendered a modal
that would periodically check whether user permissions appeared in the background.

All of this was replaced by an approach where we immediately (synchronously) ensure
that the setup is correct.